Add multi-guild support to closeAll command
Added guild awareness to closeAll command so that other servers are unchanged by the command (only the guild that the command was invoked in).

+++ src/main/kotlin/me/aberrantfox/warmbot/commands/ReportCommands.kt
@@ -32,12 +32,18 @@ fun reportCommands(reportService: ReportService, configuration: Configuration) =
                 return@execute
             }
 
+            var closeCount = 0
+
             for (report in reports) {
+                if (report.guildId != it.message.guild.id)
+                    continue
+
                 reportService.sendReportClosedEmbed(report)
                 it.jda.getTextChannelById(report.channelId).delete().queue()
+                closeCount++
             }
 
-            it.respond("All reports (${reports.size}) closed successfully.")
+            it.respond("$closeCount report(s) closed successfully.")
         }
     }
